+++ _dist/overslash-csv.js
@@ -1,5 +1,5 @@
 /** ****************************************************************************
- * overslash-csv v0.0.3
+ * overslash-csv v0.0.4
  *
  * A tiny modular Javascript utility library.
  * (you can download it from npm or github repositories)
@@ -24,6 +24,9 @@
     // only CommonJS-like environments that support module.exports,
     // like Node.
     module.exports = factory(root);
+    // This is a hack to attach the lib to the browser root when this lib is
+    // included inside another lib and the whole is browserifyied:
+    if (root.overslash === null) root.overslash = factory(root);
   } else {
     // Browser globals.
     root.overslash = factory(root);
@@ -68,7 +71,8 @@
    *  . clone             clones a literal object or an array,
    *  . extend            extends a given object with all the properties in passed-in object(s),
    *  . keys              retrieves all the names of the object's own enumerable properties,
-   *  . forPropIn         Parses all the names of the object's own enumerable properties,
+   *  . forPropIn         parses all the names of the object's own enumerable properties,
+   *  . assign            extends source with target(s) while preserving the assessors,
    *
    * Operations on Arrays (optional):
    *  . contains          returns true if the array contains the passed-in value,
@@ -157,7 +161,7 @@
   /* eslint-enable no-param-reassign */
 
   // Current version of the library:
-  overslash.VERSION = '0.0.3';
+  overslash.VERSION = '0.0.4';
 
 
   // --- Primitives types ------------------------------------------------------
@@ -472,7 +476,46 @@
           callback(key);
         }
       });
+    },
+
+    /**
+     * Extends source with target(s) while preserving the assessors.
+     *
+     * Nota:
+     * Clones a literal object at the first level while preserving the
+     * assessors (get and set). This should be the prefered method to Clones
+     * a literal object or a prototype that includes get and set assessors.
+     *
+     * Example:
+     * To clone a function prototype:
+     * var a = _.assign({}, fn.prototype);  // clone the original prototype,
+     * _.assign(fn2.prototype, a);          // assign it to fn2.prototype,
+     *
+     * @function (...arg1)
+     * @public
+     * @param {Object}      the objects to 'fusion',
+     * @returns {Object}    returns the reassigned object,
+     * @since 0.0.0
+     */
+    /* eslint-disable no-param-reassign, no-loop-func */
+    assign: function() {
+      var target = arguments[0]
+        , source
+        , descriptors
+        , i
+        ;
+
+      for (i = 1; i < arguments.length; i++) {
+        source = arguments[i];
+        descriptors = Object.keys(source).reduce(function(props, key) {
+          props[key] = Object.getOwnPropertyDescriptor(source, key);
+          return props;
+        }, {});
+      }
+      Object.defineProperties(target, descriptors);
+      return target;
     }
+    /* eslint-enable no-param-reassign, no-loop-func */
   });